MSE-004
Modificado autenticação para ficar de acordo com o que a API está retornando, criada classe para os retornos dos services, criada classe para o produto selecionado e adicionada nova rota
